Separated deck menu from rest
